+++ src/core/mod.rs
@@ -3,6 +3,7 @@ pub mod command;
 pub mod config_watcher;
 pub mod event_loop;
 pub mod input;
+pub mod lyrics_watcher;
 pub mod scheduler;
 pub mod socket;
 pub mod work;