+++ src/ui/panes/lyrics.rs
@@ -30,6 +30,16 @@ impl LyricsPane {
     pub fn new(_ctx: &Ctx) -> Self {
         Self { current_lyrics: None, initialized: false, last_requested_line_idx: 0 }
     }
+
+    fn update_lyrics(&mut self, ctx: &Ctx) -> Result<()> {
+        self.current_lyrics = None;
+
+        let lrc = ctx.find_lrc()?;
+        let Some((_, lrc)) = lrc else { return Ok(()) };
+
+        self.current_lyrics = Some(lrc);
+        Ok(())
+    }
 }
 
 impl Pane for LyricsPane {
@@ -145,14 +155,8 @@ impl Pane for LyricsPane {
 
     fn before_show(&mut self, ctx: &Ctx) -> Result<()> {
         if !self.initialized {
-            match ctx.find_lrc() {
-                Ok(lrc) => {
-                    self.current_lyrics = lrc;
-                }
-                Err(err) => {
-                    status_error!("Failed to load lyrics file: '{err}'");
-                    self.current_lyrics = None;
-                }
+            if let Err(err) = self.update_lyrics(ctx) {
+                status_error!("Failed to load lyrics file: '{err}'");
             }
             self.last_requested_line_idx = 0;
             self.initialized = true;
@@ -163,30 +167,11 @@ impl Pane for LyricsPane {
 
     fn on_event(&mut self, event: &mut UiEvent, _is_visible: bool, ctx: &Ctx) -> Result<()> {
         match event {
-            UiEvent::SongChanged | UiEvent::Reconnected => {
-                match ctx.find_lrc() {
-                    Ok(lrc) => {
-                        self.current_lyrics = lrc;
-                        ctx.render()?;
-                    }
-                    Err(err) => {
-                        self.current_lyrics = None;
-                        status_error!("Failed to load lyrics file: '{err}'");
-                    }
+            UiEvent::SongChanged | UiEvent::Reconnected | UiEvent::LyricsIndexed => {
+                if let Err(err) = self.update_lyrics(ctx) {
+                    status_error!("Failed to load lyrics file: '{err}'");
                 }
-                self.last_requested_line_idx = 0;
-            }
-            UiEvent::LyricsIndexed if self.current_lyrics.is_none() => {
-                match ctx.find_lrc() {
-                    Ok(lrc) => {
-                        self.current_lyrics = lrc;
-                        ctx.render()?;
-                    }
-                    Err(err) => {
-                        self.current_lyrics = None;
-                        status_error!("Failed to load lyrics file: '{err}'");
-                    }
-                }
+                ctx.render()?;
                 self.last_requested_line_idx = 0;
             }
             _ => {}